requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
anndata=0.8.0=pypi_0
arpack=3.7.0=hdefa2d7_2
biopython=1.78=py38h7f8727e_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.9.24=ha878542_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.0=py38h7f8727e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
contourpy=1.0.6=pypi_0
cryptography=38.0.1=py38h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
ffmpeg=4.2.2=h20bf706_0
filelock=3.6.0=pyhd3eb1b0_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h7b6447c_0
glpk=5.0=h445213a_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
h5py=3.7.0=pypi_0
huggingface_hub=0.11.1=py_0
icu=70.1=h27087fc_0
idna=3.4=py38h06a4308_0
importlib-metadata=5.1.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libhwloc=2.8.0=h32351e8_1
libiconv=1.17=h166bdaf_0
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=12.2.0=h46fd767_19
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxml2=2.10.3=h7463322_0
libzlib=1.2.13=h166bdaf_4
llvm-openmp=15.0.6=he0ac6c6_0
llvmlite=0.39.1=pypi_0
louvain=0.8.0=pypi_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.6.2=pypi_0
metis=5.1.0=h58526e2_1006
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpfr=4.1.0=h9202a9a_1
natsort=8.2.0=pypi_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
networkx=2.8.8=pypi_0
numba=0.56.4=pypi_0
numpy=1.23.4=py38h14f4228_0
numpy-base=1.23.4=py38h31eccc5_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h0b41bf4_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.2=pypi_0
patsy=0.5.3=pypi_0
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pynndescent=0.5.8=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38h06a4308_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pypi_0
python-igraph=0.10.2=py38h092921c_1
python_abi=3.8=1_cp38
pytorch=1.12.1=py3.8_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.6=pypi_0
pyyaml=6.0=py38h7f8727e_1
readline=7.0=h7b6447c_5
regex=2022.7.9=py38h5eee18b_0
requests=2.28.1=py38h06a4308_0
sacremoses=master=py_0
scanpy=1.9.1=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.9.3=pypi_0
seaborn=0.12.1=pypi_0
session-info=1.0.0=pypi_0
setuptools=65.5.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
statsmodels=0.13.5=pypi_0
stdlib-list=0.8.0=pypi_0
suitesparse=5.10.1=h9e50725_1
tbb=2021.7.0=h924138e_1
texttable=1.6.7=pyhd8ed1ab_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.10.3=py38_0
torchaudio=0.12.1=py38_cu113
torchvision=0.13.1=py38_cu113
tqdm=4.64.1=py38h06a4308_0
transformers=4.14.1=pyhd3eb1b0_0
typing-extensions=4.4.0=py38h06a4308_0
typing_extensions=4.4.0=py38h06a4308_0
umap-learn=0.5.3=pypi_0
urllib3=1.26.12=py38h06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=ha4553b6_0
